requirements.txt
django==3.2.13
requests==2.31.0
django-environ==0.9.0